Generate parsable string from PackageGenerationParameters (closes #28)
A new method is added to PackageGenerationParameters to produce a string
representation of the parameters which is parsable by the
PropertiesConfiguration.load() method.  The AutomatedPackageTool needs
to generate a stream from this string and pass it to the packager, which
reads the parameters from the stream.  The existing toString() method
does not produce a parsable string.  Because that method appears to be
used only to print debug output, it could be replaced by the new method,
if desired.