Close stream that was left open
The AutomatedPackageTool was unable to remove some of the staging area
files created during packaging because the uncompressed version of the
package was still open for reading.  This was due to the file being
opened in a constructor used as a method parameter.  That method and the
constructor parameter have been refactored into explicit calls,
including a formal stream close() call.