Close file when packaging (closes #29)
A file cleanup operation was failing in the Automated Package Tool
because a file stream remained open to the uncompressed package file in
the staging area.  This was caused by a stream constructor used as a
parameter to a copy method.  The stream is now constructed before the
copy call and closed explicitly afterward.

Modified to use try-with-resources.