Control logging output of BagItPackageAssembler for Automated Package Tool
The Automated Package Tool (APT) was printing many lines of logging
information for typical runs, which made it hard for the command line
user to see if any actual errors were being reported.  Here, most of the
logging statements are switched from "info" level to "debug" level so
that they are only seen by developers.  One message is changed to "warn"
level and another has its text edited.  The class also switches to the
ch.qos.logback.classic logger, whose level can be set dynamically to the
desired level.  The APT uses this logger and currently sets the level to
"warn", so all of the debug messages are not shown in the APT.