Insure there is only one subject represented per domain object.
A bug in URI handling led to circumstances whereby a domain object contained statements about multiple subjects.  This is not the intent; in fact, domain objects that contain multiple subjects may fail to be ingested into Fedora.

This PR fixes the bug in URI handling and insures that domain object contains statements about a single subject.