Fix Windows detection of .dotfiles
The startsWith() method for Path on Windows is unintuitive.  For a file
named ".path", startiWith(".") wil return false.  Converting to a string
and using String's startsWith() leads to the expected result.  The
consequence of this is that before this fix, Windows clients would not ignore
dotfiles when adked to do so.